refactor: move lookahead mask to generic.py
