+++ transformerx/layers/masks/__init__.py
@@ -1,3 +1,3 @@
-from .core import LookAheadMask
+from .generic import LookAheadMask
 from .core import PaddingMask
 from .core import BaseMask
